Move Config struct and related functions to pkg
